refactor: rework submit logic and errors
